task implementations

+++ src/main/java/ua/rd/pizzaservice/domain/Pizza.java
@@ -5,16 +5,13 @@ import lombok.Data;
 
 import java.math.BigDecimal;
 
-/**
- * Created by Kostiantyn_Minkov on 10/4/2016.
- */
 
 @Data
 @AllArgsConstructor
 public class Pizza {
 
-        private Integer id;
-        private String name;
-        private BigDecimal price;
-        private PizzaType type;
+    private Integer id;
+    private String name;
+    private Long price;
+    private PizzaType type;
 }
